Updated the checkbox editing journey
Highlighting in the preview follows focus in the editor. Focus in the preview follows the movements when re-ordering.
When you edit list..the add another option button is hidden and the save changes button becomes a primary one. Also when you enter edit mode the focus is moved to the first re-ordering action button in the list